Minor_ML_Project-master/requirements.txt
absl-py==0.7.1
altair==4.0.0
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.7
argh==0.26.2
arviz==0.7.0
asgiref==3.2.3
asn1crypto==1.3.0
astor==0.8.0
astroid==2.2.5
attrs==19.3.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==1.0.3
beautifulsoup4==4.8.2
bleach==3.1.0
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.10.50
botocore==1.13.50
Bottleneck==1.3.2
branca==0.4.0
breadability==0.1.20
catalogue==1.0.0
catboost==0.23
certifi==2020.4.5.1
cffi==1.14.0
cftime==1.1.1.2
chardet==3.0.4
chart-studio==1.0.0
ci-info==0.2.0
click==7.1.1
clyent==1.2.2
cmdstanpy==0.4.0
colorama==0.4.3
conda==4.8.3
conda-build==3.10.5
conda-package-handling==1.6.0
conda-verify==3.4.2
configobj==5.0.6
configparser==5.0.0
convertdate==2.2.0
cookies==2.2.1
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
Cython==0.29.16
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.3
django-crispy-forms==1.8.1
docopt==0.6.2
docutils==0.15.2
dominate==2.3.5
en-core-web-lg==2.2.5
en-core-web-sm==2.2.5
entrypoints==0.3
enum-compat==0.0.3
ephem==3.7.7.1
etelemetry==0.2.1
fastai==1.0.60
fastprogress==0.2.3
fbprophet==0.6
filelock==3.0.12
fitz==0.0.1.dev2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.18.2
gast==0.2.2
gensim==3.8.1
glob2==0.7
google-pasta==0.1.7
graphviz==0.14
grpcio==1.21.1
gunicorn==19.9.0
holidays==0.10.1
httplib2==0.18.0
idna==2.9
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
lightgbm==2.3.1
LunarCalendar==0.0.9
lxml==4.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mechanize==0.4.3
menuinst==1.4.16
mistune==0.8.4
mkl-service==2.3.0
mlxtend==0.17.2
mock==3.0.5
murmurhash==1.0.2
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
netCDF4==1.4.1
networkx==2.4
neurdflib==5.0.1
nibabel==3.1.0
nipype==1.4.2
nltk==3.4.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.1.0.25
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pathtools==0.1.2
patsy==0.5.1
pdf2image==1.13.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.0.0
pip-upgrade-outdated==1.5
pipenv==2018.11.26
pkginfo==1.5.0.1
plac==1.1.3
plotly==4.4.1
plotly-express==0.4.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.8.0
prov==1.5.3
psutil==5.7.0
pycosat==0.6.3
pycountry==19.8.18
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pygal==2.4.0
Pygments==2.6.1
PyInstaller==3.6
PyMeeus==0.3.7
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyrsistent==0.16.0
PySocks==1.7.1
pystan==2.19.1.1
pytesseract==0.3.1
python-dateutil==2.8.0
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyxnat==1.3
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
rdflib==5.0.0
requests==2.23.0
retrying==1.3.3
ruamel-yaml==0.15.87
s3transfer==0.2.1
scikit-learn==0.22.2.post1
scipy==1.3.1
seaborn==0.10.0
Send2Trash==1.5.0
setuptools-git==1.2
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smart-open==1.9.0
soupsieve==2.0
spacy==2.2.3
sqlparse==0.3.0
srsly==1.0.1
statsmodels==0.11.1
streamlit==0.52.2
sumy==0.8.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
tesseract==0.1.3
testpath==0.4.4
textblob==0.15.3
Theano==1.0.4
thinc==7.3.1
toml==0.10.0
toolz==0.10.0
torch==1.1.0
torchvision==0.1.8
tornado==5.1.1
tqdm==4.44.1
traitlets==4.3.3
traits==6.0.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.1
virtualenv==16.7.9
virtualenv-clone==0.5.3
visitor==0.1.3
wasabi==0.6.0
watchdog==0.9.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.6.0
wrapt==1.11.2
WTForms==2.2.1
WTForms-SQLAlchemy==0.1
xarray==0.15.1
xgboost==0.90
zipp==2.2.0
